% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pcr.R
\name{plot_pcr}
\alias{plot_pcr}
\title{Plot PCR Data}
\usage{
plot_pcr(pcr_data, Ct)
}
\arguments{
\item{pcr_data}{A data frame with PCR data (as output from \code{simulate_pcr}).}

\item{Ct}{The cycle threshold value (as output from \code{simulate_pcr}).}
}
\description{
This function plots the PCR fluorescence data and the cycle threshold.
}
