% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pcr.R
\name{simulate_pcr}
\alias{simulate_pcr}
\title{Simulate PCR Data}
\usage{
simulate_pcr(N0, E, threshold = 10000)
}
\arguments{
\item{N0}{Initial amount of DNA.}

\item{E}{Efficiency of the PCR process.}

\item{threshold}{Threshold value for detecting the cycle.}
}
\value{
A list with two elements: pcr_data (a data frame) and Ct (the cycle threshold).
}
\description{
This function simulates PCR fluorescence data based on initial DNA amount,
efficiency, and a threshold value.
}
